requirements-web.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
SQLAlchemy==2.0.*
PyJWT==2.9.*
python-dotenv==1.0.*
